requirements.txt
alembic==1.8.1
SQLAlchemy==1.4.42
fastapi==0.85.1
pydantic==1.10.2
python-jose==3.3.0
cryptography==38.0.1
passlib==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.4
tenacity==8.1.0
ipython==8.5.0
email-validator==1.3.0
python-multipart==0.0.5
pytest==7.1.3
sqlalchemy-utils==0.38.3